requirements.txt
gpxpy==1.6.2
lxml==5.2.2
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
